Move magick helpers to Utilities
Use EndTransaction to batch changes
